commands/globals: wrap apply in an error handler
As we split the apply method up into separate methods we need a way for
those helpers to stop execution of the method. This wraps the real apply
method inside of a try/except block with a custom exception. Any helper
that is intended to run in the apply method simply needs to raise this
exception, and apply will return.